backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
playwright==1.40.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
rapidfuzz==3.5.2
pandas==2.1.4
numpy==1.25.2
